renamed to OpenTelemetry.Sampler.AWS
