+++ src/OpenTelemetry.Sampler.AWS/.publicApi/net462/PublicAPI.Shipped.txt
