+++ src/OpenTelemetry.Sampler.AWS/.publicApi/net6.0/PublicAPI.Shipped.txt
