+++ src/OpenTelemetry.Sampler.AWS/.publicApi/netstandard2.0/PublicAPI.Shipped.txt
