+++ src/OpenTelemetry.Sampler.AWS/CHANGELOG.md
@@ -2,5 +2,7 @@
 
 ## Unreleased
 
+Initial release of `OpenTelemetry.Samplers.AWS`.
+
 * Feature - AWSXRayRemoteSampler - Add support for AWS X-Ray remote sampling
   ([#1091](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1091))
