+++ src/OpenTelemetry.Sampler.AWS/GetSamplingRulesResponse.cs
@@ -1,4 +1,4 @@
-// <copyright file="AssemblyInfo.cs" company="OpenTelemetry Authors">
+// <copyright file="GetSamplingRulesResponse.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,10 +13,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
-using System.Runtime.CompilerServices;
 
-#if SIGNED
-[assembly: InternalsVisibleTo("OpenTelemetry.Samplers.AWS.Tests, PublicKey=002400000480000094000000060200000024000052534131000400000100010051c1562a090fb0c9f391012a32198b5e5d9a60e9b80fa2d7b434c9e5ccb7259bd606e66f9660676afc6692b8cdc6793d190904551d2103b7b22fa636dcbb8208839785ba402ea08fc00c8f1500ccef28bbf599aa64ffb1e1d5dc1bf3420a3777badfe697856e9d52070a50c3ea5821c80bef17ca3acffa28f89dd413f096f898")]
-#else
-[assembly: InternalsVisibleTo("OpenTelemetry.Samplers.AWS.Tests")]
-#endif
+using System.Collections.Generic;
+using System.Text.Json.Serialization;
+
+namespace OpenTelemetry.Sampler.AWS;
+
+internal class GetSamplingRulesResponse
+{
+    [JsonPropertyName("NextToken")]
+    public string? NextToken { get; set; }
+
+    [JsonPropertyName("SamplingRuleRecords")]
+    public List<SamplingRuleRecord>? SamplingRuleRecords { get; set; }
+}
