+++ src/OpenTelemetry.Sampler.AWS/SamplingRuleRecord.cs
@@ -16,7 +16,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace OpenTelemetry.Samplers.AWS;
+namespace OpenTelemetry.Sampler.AWS;
 
 internal class SamplingRuleRecord
 {
