+++ test/OpenTelemetry.Sampler.AWS.Tests/Data/GetSamplingRulesResponse.json
